許願池下架後台時間排程完成 發布者資訊完成 顯示剩餘需求數量和已募集數量完成

+++ EngeeGitHub/EnGeeMVC/EnGee/Controllers/CollectController.cs
@@ -91,8 +91,12 @@ namespace EnGee.Controllers
                      MainCategoryId = co.MainCategoryId,
                      SubcategoryId = co.SubcategoryId,
                      CollectStatus = co.CollectStatus,
+                     member = m,
+                     TotalPublished = db.TCollects
+                        .Where(c => c.MemberId == m.MemberId && c.CollectStatus == true)
+                        .Count(),
                  };
-
+            
             // 日期排序
             List<Rong_CollectIndexViewModel> orderedCollectIndex = collectindex.ToList();
             if (sortBy == 1)
@@ -178,9 +182,27 @@ namespace EnGee.Controllers
                      ConvenienNum = c.ConvenienNum != null ? c.ConvenienNum : "",
                      CollectImagePath = c.CollectImagePath,
                      CollectItemName = c.CollectItemName,
-                     CollectAmount = c.CollectAmount
+                     CollectAmount = c.CollectAmount,
+                     member = m,  //   取得發布者的資料
                  })
                 .FirstOrDefault();
+
+            //  目前已徵求到的數量
+            int TotalDonationAmount = 0;
+            TotalDonationAmount = db.TDonationOrders
+                .Where(d => d.CollectId == id)
+                .Sum(d => d.DonationAmount);
+            ViewBag.TotalDonationAmount = TotalDonationAmount;
+
+            
+
+            //  該會員共發布幾篇許願
+            int TotalPublished = 0;
+            TotalPublished = db.TCollects
+                .Where(c => c.MemberId == cinfo.MemberId)
+                .Count();
+            ViewBag.Published = TotalPublished;
+
 
             // 登入判斷
             if (!HttpContext.Session.Keys.Contains(CDictionary.SK_LOINGED_USER))
@@ -287,8 +309,6 @@ namespace EnGee.Controllers
         // 會員_許願刪除
         public IActionResult DeleteCollectMember(int? id)
         {
-           
-
             if (id == null)
                 return RedirectToAction("Create");
 
@@ -321,7 +341,7 @@ namespace EnGee.Controllers
             {
                 TMember loggedInUser = JsonSerializer.Deserialize<TMember>(userJson);
 
-                if (loggedInUser.MemberId != c.MemberId)
+                if (loggedInUser.Access == 1 || (loggedInUser.Access == 3 && loggedInUser.MemberId != c.MemberId))
                 {
                     return RedirectToAction("CollectIndex");
                 }
@@ -429,7 +449,7 @@ namespace EnGee.Controllers
                 ViewBag.MemberPoint = loggedInUser.Point;
             }
 
-            // 扣除點數
+            // 扣除點數 + 扣除徵求數量
             if (loggedInUser != null && collect != null && ViewBag.MemberId != null)
             {
                 var selectedDeliveryType = db.TDeliveryTypes.FirstOrDefault(dt => dt.DeliveryTypeId == d.DeliveryTypeId);
@@ -520,18 +540,19 @@ namespace EnGee.Controllers
             {
                 TMember loggedInUser = JsonSerializer.Deserialize<TMember>(userJson);
 
-                if (loggedInUser.Access != 0 || loggedInUser.Access != 2)
+                if (loggedInUser.Access == 1 || loggedInUser.Access == 3)
                 {
                     return RedirectToAction("CollectIndex");
                 }
+                else
+                {
+                    if (id == null)
+                        return RedirectToAction("DonationManagement");
+                }
             }
-
-            if (id == null)
-                return RedirectToAction("DonationManagement");
-
             TDonationOrder d = db.TDonationOrders
-                .Include(t => t.Collect)
-                .FirstOrDefault(t => t.DonationOrderId == id);
+                        .Include(t => t.Collect)
+                        .FirstOrDefault(t => t.DonationOrderId == id);
 
             if (d == null)
                 return RedirectToAction("DonationManagement");
@@ -545,18 +566,47 @@ namespace EnGee.Controllers
         [HttpPost]
         public IActionResult EditDonation(Rong_CDonationWrap dIn)
         {
-            TDonationOrder dDb = db.TDonationOrders.FirstOrDefault(t => t.DonationOrderId == dIn.DonationOrderId);
+            string userJson = HttpContext.Session.GetString(CDictionary.SK_LOINGED_USER);
+            if (userJson == null)
+                return RedirectToAction("Login", "Home");
 
-            if (dDb != null)
+            TMember loggedInUser = JsonSerializer.Deserialize<TMember>(userJson);
+
+            if (loggedInUser.Access == 0 || loggedInUser.Access == 2)
             {
-                dDb.DonarName = dIn.DonarName;
-                dDb.DonarPhone = dIn.DonarPhone;
-                dDb.DonationAmount = dIn.DonationAmount;
-                dDb.DonationStatus = dIn.DonationStatus;
-                db.SaveChanges();
+
+                TDonationOrder dDb = db.TDonationOrders.FirstOrDefault(t => t.DonationOrderId == dIn.DonationOrderId);
+
+                if (dDb != null)
+                {   
+                    //  把修改的捐贈數量加減回徵求數量
+                    var collect = db.TCollects.FirstOrDefault(c => c.CollectId == dDb.CollectId);
+                    if (collect != null) 
+                    {
+                        if (dIn.DonationAmount > dDb.DonationAmount) 
+                        {
+                            int newCollectAmount = collect.CollectAmount - (dIn.DonationAmount - dDb.DonationAmount);
+                        }
+                        else if (dIn.DonationAmount < dDb.DonationAmount)
+                        {
+                            int newCollectAmount = collect.CollectAmount + (dDb.DonationAmount - dIn.DonationAmount);
+                            collect.CollectAmount = newCollectAmount;
+                        }
+                    }
+                    dDb.DonarName = dIn.DonarName;
+                    dDb.DonarPhone = dIn.DonarPhone;
+                    dDb.DonationAmount = dIn.DonationAmount;
+                    dDb.DonationStatus = dIn.DonationStatus;
+                    db.SaveChanges();
+                }
+
+                return RedirectToAction("DonationManagement");
             }
 
-            return RedirectToAction("DonationManagement");
+            else
+            {
+                return RedirectToAction("CollectIndex");
+            }
         }
 
         // 捐贈刪除
@@ -568,13 +618,22 @@ namespace EnGee.Controllers
             TDonationOrder d = db.TDonationOrders.FirstOrDefault(t => t.DonationOrderId == id);
             if (d != null)
             {
+                //  把刪除的捐贈數量加回徵求數量
                 var collect = db.TCollects.FirstOrDefault(c => c.CollectId == d.CollectId);
 
-                if (collect != null) 
+                //  把刪除的運費加回會員點數
+                var selectedDeliveryType = db.TDeliveryTypes.FirstOrDefault(dt => dt.DeliveryTypeId == d.DeliveryTypeId);
+                var member = db.TMembers.FirstOrDefault(m => m.MemberId == d.MemberId);
+
+                if (collect != null && selectedDeliveryType != null) 
                 {
                     int newCollectAmount = collect.CollectAmount + d.DonationAmount;
-                    Console.WriteLine(d.DonationAmount.ToString());
                     collect.CollectAmount = newCollectAmount;
+
+                    int deliveryFee = (int)selectedDeliveryType.DeliveryFee;
+                    int newPoints = (int)member.Point + deliveryFee;
+                    member.Point = newPoints;
+
                     db.TDonationOrders.Remove(d);
                     db.SaveChanges();
                 }
@@ -582,5 +641,7 @@ namespace EnGee.Controllers
             }
             return RedirectToAction("DonationManagement");
         }
+
+        
     }
 }
